+++ packages/db/src/schema/ap-exams.ts
@@ -1,0 +1,46 @@
+import { integer, json, pgTable, uniqueIndex, uuid, varchar } from "drizzle-orm/pg-core";
+
+export type APCoursesGrantedTree =
+  | {
+      AND: (APCoursesGrantedTree | string)[];
+    }
+  | {
+      OR: (APCoursesGrantedTree | string)[];
+    };
+
+export const apExam = pgTable("ap_exam", {
+  id: varchar("id").primaryKey(),
+  catalogueName: varchar("catalogue_name"),
+});
+
+export const apExamToReward = pgTable(
+  "ap_exam_to_reward",
+  {
+    id: uuid("id").primaryKey().defaultRandom(),
+    examId: varchar("exam_id")
+      .notNull()
+      .references(() => apExam.id, { onDelete: "cascade" }),
+    score: integer("score").notNull(),
+    reward: uuid("reward")
+      .notNull()
+      .references(() => apExamReward.id, { onDelete: "cascade" }),
+  },
+  (table) => [uniqueIndex().on(table.examId, table.score)],
+);
+
+export const apExamReward = pgTable("ap_exam_reward", {
+  id: uuid("id").primaryKey().defaultRandom(),
+  unitsGranted: integer("units_granted").notNull(),
+  electiveUnitsGranted: integer("elective_units_granted").notNull(),
+  ge1aCoursesGranted: integer("ge_1a_courses_granted").notNull().default(0),
+  ge1bCoursesGranted: integer("ge_1b_courses_granted").notNull().default(0),
+  ge2CoursesGranted: integer("ge_2_courses_granted").notNull().default(0),
+  ge3CoursesGranted: integer("ge_3_courses_granted").notNull().default(0),
+  ge4CoursesGranted: integer("ge_4_courses_granted").notNull().default(0),
+  ge5aCoursesGranted: integer("ge_5a_courses_granted").notNull().default(0),
+  ge5bCoursesGranted: integer("ge_5b_courses_granted").notNull().default(0),
+  ge6CoursesGranted: integer("ge_6_courses_granted").notNull().default(0),
+  ge7CoursesGranted: integer("ge_7_courses_granted").notNull().default(0),
+  ge8CoursesGranted: integer("ge_8_courses_granted").notNull().default(0),
+  coursesGranted: json("courses_granted").$type<APCoursesGrantedTree>().notNull(),
+});
