+++ packages/db/src/schema/instructors.ts
@@ -1,0 +1,130 @@
+import { eq, getTableColumns, ne, sql } from "drizzle-orm";
+import {
+  index,
+  pgMaterializedView,
+  pgTable,
+  uniqueIndex,
+  uuid,
+  varchar,
+} from "drizzle-orm/pg-core";
+import { courses } from "./courses";
+import { websocCourse, websocInstructor, websocSection, websocSectionToInstructor } from "./websoc";
+
+export const instructor = pgTable(
+  "instructor",
+  {
+    ucinetid: varchar("ucinetid").primaryKey(),
+    name: varchar("name").notNull(),
+    title: varchar("title").notNull(),
+    email: varchar("email").notNull(),
+    department: varchar("department").notNull(),
+  },
+  (table) => [
+    index("instructor_search_index").using(
+      "gin",
+      sql`(
+ SETWEIGHT(TO_TSVECTOR('english', COALESCE(${table.ucinetid}, '')), 'A') ||
+ SETWEIGHT(TO_TSVECTOR('english', COALESCE(${table.name}, '')), 'B') ||
+ SETWEIGHT(TO_TSVECTOR('english', COALESCE(${table.title}, '')), 'B')
+)`,
+    ),
+  ],
+);
+
+export const instructorToWebsocInstructor = pgTable(
+  "instructor_to_websoc_instructor",
+  {
+    id: uuid("id").primaryKey().defaultRandom(),
+    instructorUcinetid: varchar("instructor_ucinetid").references(() => instructor.ucinetid),
+    websocInstructorName: varchar("websoc_instructor_name")
+      .notNull()
+      .references(() => websocInstructor.name),
+  },
+  (table) => [
+    index().on(table.instructorUcinetid),
+    index().on(table.websocInstructorName),
+    uniqueIndex().on(table.instructorUcinetid, table.websocInstructorName),
+  ],
+);
+
+export const instructorView = pgMaterializedView("instructor_view").as((qb) => {
+  const shortenedNamesCte = qb.$with("shortened_names_cte").as(
+    qb
+      .select({
+        instructorUcinetid: instructorToWebsocInstructor.instructorUcinetid,
+        shortenedNames: sql`ARRAY_AGG(${instructorToWebsocInstructor.websocInstructorName})`.as(
+          "shortened_names",
+        ),
+      })
+      .from(instructorToWebsocInstructor)
+      .groupBy(instructorToWebsocInstructor.instructorUcinetid),
+  );
+  const termsCte = qb.$with("terms_cte").as(
+    qb
+      .select({
+        courseId: courses.id,
+        instructorUcinetid: instructorToWebsocInstructor.instructorUcinetid,
+        terms: sql`
+          ARRAY_REMOVE(ARRAY_AGG(DISTINCT
+            CASE WHEN ${websocCourse.year} IS NULL THEN NULL
+            ELSE CONCAT(${websocCourse.year}, ' ', ${websocCourse.quarter})
+            END
+          ), NULL)`.as("terms"),
+      })
+      .from(courses)
+      .leftJoin(websocCourse, eq(websocCourse.courseId, courses.id))
+      .leftJoin(websocSection, eq(websocSection.courseId, websocCourse.id))
+      .leftJoin(
+        websocSectionToInstructor,
+        eq(websocSectionToInstructor.sectionId, websocSection.id),
+      )
+      .leftJoin(
+        websocInstructor,
+        eq(websocInstructor.name, websocSectionToInstructor.instructorName),
+      )
+      .leftJoin(
+        instructorToWebsocInstructor,
+        eq(instructorToWebsocInstructor.websocInstructorName, websocInstructor.name),
+      )
+      .groupBy(courses.id, instructorToWebsocInstructor.instructorUcinetid),
+  );
+  const coursesCte = qb.$with("courses_cte").as(
+    qb
+      .with(termsCte)
+      .select({
+        instructorUcinetid: termsCte.instructorUcinetid,
+        courseId: courses.id,
+        courseInfo: sql`
+          CASE WHEN ${courses.id} IS NULL
+          THEN NULL
+          ELSE JSONB_BUILD_OBJECT(
+               'id', ${courses.id},
+               'title', ${courses.title},
+               'department', ${courses.department},
+               'courseNumber', ${courses.courseNumber},
+               'terms', COALESCE(${termsCte.terms}, ARRAY[]::TEXT[])
+          )
+          END
+          `.as("course_info"),
+      })
+      .from(courses)
+      .leftJoin(termsCte, eq(termsCte.courseId, courses.id))
+      .groupBy(courses.id, termsCte.instructorUcinetid, termsCte.terms),
+  );
+  return qb
+    .with(shortenedNamesCte, coursesCte)
+    .select({
+      ...getTableColumns(instructor),
+      shortenedNames: sql<
+        string[]
+      >`COALESCE(${shortenedNamesCte.shortenedNames}, ARRAY[]::TEXT[])`.as("shortened_names"),
+      courses: sql`
+          ARRAY_REMOVE(ARRAY_AGG(DISTINCT ${coursesCte.courseInfo}), NULL)
+        `.as("courses"),
+    })
+    .from(instructor)
+    .leftJoin(shortenedNamesCte, eq(shortenedNamesCte.instructorUcinetid, instructor.ucinetid))
+    .leftJoin(coursesCte, eq(coursesCte.instructorUcinetid, instructor.ucinetid))
+    .where(ne(instructor.ucinetid, "student"))
+    .groupBy(instructor.ucinetid, shortenedNamesCte.shortenedNames);
+});
